feat: FAQs page created with seperated components and Toast(Sooner) added
